+++ cmd/blockchaincmd/convert.go
@@ -617,7 +617,7 @@ func convertBlockchain(_ *cobra.Command, args []string) error {
 			}
 
 		default:
-			bootstrapValidators, _, err = promptBootstrapValidators(
+			bootstrapValidators, err = promptBootstrapValidators(
 				network,
 				changeOwnerAddress,
 				numBootstrapValidators,
